requirements.txt
streamlit
PyPDF2
python-docx
langdetect
spacy
openai
